requirements.txt
pyserial >=3.4
pytest >=4.5.0
pyusb >=1.0.2
